+++ docs/html/annotated_dup.js
@@ -1,4 +1,6 @@
 var annotated_dup =
 [
+    [ "_p2p_settings", "d9/db2/struct__p2p__settings.html", "d9/db2/struct__p2p__settings" ],
+    [ "_stParam", "d7/d6b/struct__st_param.html", "d7/d6b/struct__st_param" ],
     [ "RUI3", "d2/ded/class_r_u_i3.html", "d2/ded/class_r_u_i3" ]
 ];