+++ docs/html/d0/dd0/rui3__at_8h.js
@@ -1,5 +1,7 @@
 var rui3__at_8h =
 [
+    [ "_stParam", "d7/d6b/struct__st_param.html", "d7/d6b/struct__st_param" ],
+    [ "_p2p_settings", "d9/db2/struct__p2p__settings.html", "d9/db2/struct__p2p__settings" ],
     [ "RUI3", "d2/ded/class_r_u_i3.html", "d2/ded/class_r_u_i3" ],
     [ "ABP", "d0/dd0/rui3__at_8h.html#a9075988bf7c71939dfd0bee9a4d656d9", null ],
     [ "ARRAY_SIZE", "d0/dd0/rui3__at_8h.html#a70c57aae3eb654e205459b4362c8089a", null ],
@@ -25,9 +27,13 @@ var rui3__at_8h =
     [ "LPM_LVL_2", "d0/dd0/rui3__at_8h.html#ab3bc7b7923cbeb37a40ffca28411d68a", null ],
     [ "LPM_OFF", "d0/dd0/rui3__at_8h.html#a29e6c9edc1e565eaacad0e3a060e380e", null ],
     [ "LPM_ON", "d0/dd0/rui3__at_8h.html#a38eca721416840b7a6ce26536d3aa116", null ],
+    [ "MAX_ARGUMENT", "d0/dd0/rui3__at_8h.html#a2ca4a055e57a59da428f7b808d83611d", null ],
+    [ "MAX_CMD_LEN", "d0/dd0/rui3__at_8h.html#a1eb73c104b484cf18752169509cebfe2", null ],
     [ "NO_RESPONSE", "d0/dd0/rui3__at_8h.html#a5f137187c6a892935c5acd036f609b22", null ],
     [ "OTAA", "d0/dd0/rui3__at_8h.html#a2b37538ab3be04ba4ee0779aaaf2857d", null ],
     [ "RU864", "d0/dd0/rui3__at_8h.html#a8478d6919485d749d55c2cf82ae90766", null ],
     [ "UNCONF", "d0/dd0/rui3__at_8h.html#a76648e4c1862c605e92ad0f3609c316a", null ],
-    [ "US915", "d0/dd0/rui3__at_8h.html#a1c3ceda1abdbc55ab9b9df3616632769", null ]
+    [ "US915", "d0/dd0/rui3__at_8h.html#a1c3ceda1abdbc55ab9b9df3616632769", null ],
+    [ "p2p_settings", "d0/dd0/rui3__at_8h.html#a2db711d8bb7f9ecffd3f641421729412", null ],
+    [ "stParam", "d0/dd0/rui3__at_8h.html#aa581028d3a930b08bfad9f01581799e8", null ]
 ];