+++ docs/html/search/all_3.js
@@ -1,5 +1,10 @@
 var searchData=
 [
-  ['eu433_0',['EU433',['../d0/dd0/rui3__at_8h.html#a8dea9d11de184009595db420d3015e44',1,'rui3_at.h']]],
-  ['eu868_1',['EU868',['../d0/dd0/rui3__at_8h.html#aea0f54c3f5cc44ec2fd9dc532c27c8a0',1,'rui3_at.h']]]
+  ['class_5fa_0',['CLASS_A',['../d0/dd0/rui3__at_8h.html#a95f496e7b10a8915b6dfdb499571df20',1,'rui3_at.h']]],
+  ['class_5fb_1',['CLASS_B',['../d0/dd0/rui3__at_8h.html#abb7d635bd987ddd07ac47b37e8e96828',1,'rui3_at.h']]],
+  ['class_5fc_2',['CLASS_C',['../d0/dd0/rui3__at_8h.html#aeb5406b68f24386e785702b84ce1d091',1,'rui3_at.h']]],
+  ['cn470_3',['CN470',['../d0/dd0/rui3__at_8h.html#a2f0d9f210d7488de6dfac3b85277910b',1,'rui3_at.h']]],
+  ['command_4',['command',['../de/d6d/rui3__at_8cpp.html#a09b945691015281174f3851d8470ec7f',1,'rui3_at.cpp']]],
+  ['conf_5',['CONF',['../d0/dd0/rui3__at_8h.html#afb6b8c4d21f9183c9c65f548d976af2f',1,'rui3_at.h']]],
+  ['cr_6',['cr',['../d9/db2/struct__p2p__settings.html#a239a9eff6291cb5ef6129606416b8d32',1,'_p2p_settings']]]
 ];