+++ docs/html/search/all_4.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['flushrx_0',['flushRX',['../d2/ded/class_r_u_i3.html#a8462fda0a7618a74da3504652820e1ba',1,'RUI3']]]
+  ['eu433_0',['EU433',['../d0/dd0/rui3__at_8h.html#a8dea9d11de184009595db420d3015e44',1,'rui3_at.h']]],
+  ['eu868_1',['EU868',['../d0/dd0/rui3__at_8h.html#aea0f54c3f5cc44ec2fd9dc532c27c8a0',1,'rui3_at.h']]]
 ];