+++ docs/html/search/all_8.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['kr920_0',['KR920',['../d0/dd0/rui3__at_8h.html#ab240a34706b9bec388ce50a86f248f01',1,'rui3_at.h']]]
+  ['joinloranetwork_0',['joinLoRaNetwork',['../d2/ded/class_r_u_i3.html#a7f839a1e56f928020ff28f3709973fb7',1,'RUI3']]]
 ];