+++ docs/html/search/all_9.js
@@ -1,11 +1,4 @@
 var searchData=
 [
-  ['la915_0',['LA915',['../d0/dd0/rui3__at_8h.html#a719efbfefb433f4864251e27d286898d',1,'rui3_at.h']]],
-  ['list_1',['Todo List',['../dd/da0/todo.html',1,'']]],
-  ['lorap2p_2',['LoRaP2P',['../d0/dd0/rui3__at_8h.html#a3083a5c58e5626829bb0c4594ece85c7',1,'rui3_at.h']]],
-  ['lorawan_3',['LoRaWAN',['../d0/dd0/rui3__at_8h.html#aabd0258c9a3de8af9717b9336b66859b',1,'rui3_at.h']]],
-  ['lpm_5flvl_5f1_4',['LPM_LVL_1',['../d0/dd0/rui3__at_8h.html#a6ca89ecb1eb33cf16c48a90111bac7d4',1,'rui3_at.h']]],
-  ['lpm_5flvl_5f2_5',['LPM_LVL_2',['../d0/dd0/rui3__at_8h.html#ab3bc7b7923cbeb37a40ffca28411d68a',1,'rui3_at.h']]],
-  ['lpm_5foff_6',['LPM_OFF',['../d0/dd0/rui3__at_8h.html#a29e6c9edc1e565eaacad0e3a060e380e',1,'rui3_at.h']]],
-  ['lpm_5fon_7',['LPM_ON',['../d0/dd0/rui3__at_8h.html#a38eca721416840b7a6ce26536d3aa116',1,'rui3_at.h']]]
+  ['kr920_0',['KR920',['../d0/dd0/rui3__at_8h.html#ab240a34706b9bec388ce50a86f248f01',1,'rui3_at.h']]]
 ];