+++ docs/html/search/all_b.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['otaa_0',['OTAA',['../d0/dd0/rui3__at_8h.html#a2b37538ab3be04ba4ee0779aaaf2857d',1,'rui3_at.h']]]
+  ['max_5fargument_0',['MAX_ARGUMENT',['../d0/dd0/rui3__at_8h.html#a2ca4a055e57a59da428f7b808d83611d',1,'rui3_at.h']]],
+  ['max_5fcmd_5flen_1',['MAX_CMD_LEN',['../d0/dd0/rui3__at_8h.html#a1eb73c104b484cf18752169509cebfe2',1,'rui3_at.h']]]
 ];