+++ docs/html/search/classes_0.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['rui3_0',['RUI3',['../d2/ded/class_r_u_i3.html',1,'']]]
+  ['_5fp2p_5fsettings_0',['_p2p_settings',['../d9/db2/struct__p2p__settings.html',1,'']]],
+  ['_5fstparam_1',['_stParam',['../d7/d6b/struct__st_param.html',1,'']]]
 ];