+++ docs/html/search/defines_6.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['no_5fresponse_0',['NO_RESPONSE',['../d0/dd0/rui3__at_8h.html#a5f137187c6a892935c5acd036f609b22',1,'rui3_at.h']]]
+  ['max_5fargument_0',['MAX_ARGUMENT',['../d0/dd0/rui3__at_8h.html#a2ca4a055e57a59da428f7b808d83611d',1,'rui3_at.h']]],
+  ['max_5fcmd_5flen_1',['MAX_CMD_LEN',['../d0/dd0/rui3__at_8h.html#a1eb73c104b484cf18752169509cebfe2',1,'rui3_at.h']]]
 ];