+++ docs/html/search/defines_7.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['otaa_0',['OTAA',['../d0/dd0/rui3__at_8h.html#a2b37538ab3be04ba4ee0779aaaf2857d',1,'rui3_at.h']]]
+  ['no_5fresponse_0',['NO_RESPONSE',['../d0/dd0/rui3__at_8h.html#a5f137187c6a892935c5acd036f609b22',1,'rui3_at.h']]]
 ];