+++ docs/html/search/defines_8.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['ru864_0',['RU864',['../d0/dd0/rui3__at_8h.html#a8478d6919485d749d55c2cf82ae90766',1,'rui3_at.h']]]
+  ['otaa_0',['OTAA',['../d0/dd0/rui3__at_8h.html#a2b37538ab3be04ba4ee0779aaaf2857d',1,'rui3_at.h']]]
 ];