+++ docs/html/search/functions_4.js
@@ -2,5 +2,5 @@ var searchData=
 [
   ['initabp_0',['initABP',['../d2/ded/class_r_u_i3.html#a0d43db6f04785e2519c856a802e6e062',1,'RUI3']]],
   ['initotaa_1',['initOTAA',['../d2/ded/class_r_u_i3.html#a15d6925f150158dce908ebf6040a35fd',1,'RUI3']]],
-  ['initp2p_2',['initP2P',['../d2/ded/class_r_u_i3.html#a341914b651ad57d5fd3f5be1ac3c873c',1,'RUI3']]]
+  ['initp2p_2',['initP2P',['../d2/ded/class_r_u_i3.html#aa8fb2364278d16fa99dd265589aa8e95',1,'RUI3']]]
 ];