+++ docs/html/search/functions_7.js
@@ -9,8 +9,9 @@ var searchData=
   ['setjoinmode_6',['setJoinMode',['../d2/ded/class_r_u_i3.html#a47948e4a085c35c575c6fab33f16c929',1,'RUI3']]],
   ['setlpm_7',['setLPM',['../d2/ded/class_r_u_i3.html#abea6971995336b4a510cac6e4e75d2e0',1,'RUI3']]],
   ['setlpmlevel_8',['setLPMLevel',['../d2/ded/class_r_u_i3.html#a82cdf3e93f8825375150647fb3a32d08',1,'RUI3']]],
-  ['setregion_9',['setRegion',['../d2/ded/class_r_u_i3.html#a073fb5f2e8adc83de8b806e8dfda6fb5',1,'RUI3']]],
-  ['setuartconfig_10',['setUARTConfig',['../d2/ded/class_r_u_i3.html#a68de628dd5ac6b2938c93e23710e1657',1,'RUI3']]],
-  ['setworkingmode_11',['setWorkingMode',['../d2/ded/class_r_u_i3.html#ac627738073661f4d8f2a9ef6b2e79b7d',1,'RUI3']]],
-  ['sleep_12',['sleep',['../d2/ded/class_r_u_i3.html#a59241e569d25b8ecff616ef4cd5a5d6b',1,'RUI3']]]
+  ['setp2pcad_9',['setP2PCAD',['../d2/ded/class_r_u_i3.html#a10441c2001f38f573e62a33b45bcaee0',1,'RUI3']]],
+  ['setregion_10',['setRegion',['../d2/ded/class_r_u_i3.html#a073fb5f2e8adc83de8b806e8dfda6fb5',1,'RUI3']]],
+  ['setuartconfig_11',['setUARTConfig',['../d2/ded/class_r_u_i3.html#a68de628dd5ac6b2938c93e23710e1657',1,'RUI3']]],
+  ['setworkingmode_12',['setWorkingMode',['../d2/ded/class_r_u_i3.html#ac627738073661f4d8f2a9ef6b2e79b7d',1,'RUI3']]],
+  ['sleep_13',['sleep',['../d2/ded/class_r_u_i3.html#a59241e569d25b8ecff616ef4cd5a5d6b',1,'RUI3']]]
 ];