+++ docs/html/search/searchdata.js
@@ -1,12 +1,13 @@
 var indexSectionsWithContent =
 {
-  0: "abcefgijklnorstu",
-  1: "r",
+  0: "_abcefgijklmnoprstu",
+  1: "_r",
   2: "r",
   3: "abfgijrs",
-  4: "cr",
-  5: "aceiklnoru",
-  6: "lt"
+  4: "abcfprst",
+  5: "ps",
+  6: "aceiklmnoru",
+  7: "lt"
 };
 
 var indexSectionNames =
@@ -16,8 +17,9 @@ var indexSectionNames =
   2: "files",
   3: "functions",
   4: "variables",
-  5: "defines",
-  6: "pages"
+  5: "typedefs",
+  6: "defines",
+  7: "pages"
 };
 
 var indexSectionLabels =
@@ -27,7 +29,8 @@ var indexSectionLabels =
   2: "Files",
   3: "Functions",
   4: "Variables",
-  5: "Macros",
-  6: "Pages"
+  5: "Typedefs",
+  6: "Macros",
+  7: "Pages"
 };