+++ docs/html/search/variables_6.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['sf_0',['sf',['../d9/db2/struct__p2p__settings.html#af13284de3f688b86ceafd66b9aca9076',1,'_p2p_settings']]]
+];