+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## V0.0.3 optimizations
+- Make initP2P simpler to use
+- Add P2P setter/getter functions
+
 ## V0.0.2 fix rui3_at.h
 - Fix typo in function description